+++ tests/unit/BusinessNumber.spec.ts
@@ -2,7 +2,7 @@ import Vue from 'vue'
 import flushPromises from 'flush-promises'
 import { Wrapper } from '@vue/test-utils'
 import { getLastEvent } from '../get-last-event'
-import { wrapperFactory } from '../jest-wrapper-factory'
+import { wrapperFactory } from '../vitest-wrapper-factory'
 import BusinessNumber from '@/components/Registration/BusinessNumber.vue'
 
 const updateBusinessNumberEvent = 'update:businessNumber'
