+++ tests/unit/CareAndCustodySelect.spec.ts
@@ -1,5 +1,5 @@
 import Vue from 'vue'
-import { wrapperFactory } from '../jest-wrapper-factory'
+import { wrapperFactory } from '../vitest-wrapper-factory'
 import CareAndCustodySelect from '@/components/Dissolution/CareAndCustodySelect.vue'
 import { DissolutionResources } from '@/resources'
 import { RoleTypes } from '@/enums'
