+++ tests/unit/DissolutionStatement.spec.ts
@@ -1,5 +1,5 @@
 import Vue from 'vue'
-import { wrapperFactory } from '../jest-wrapper-factory'
+import { wrapperFactory } from '../vitest-wrapper-factory'
 import DissolutionStatement from '@/components/Dissolution/DissolutionStatement.vue'
 import { DissolutionStatementTypes } from '@/enums'
 import { CorpTypeCd } from '@bcrs-shared-components/enums/'
