+++ tests/unit/EntityInfo.spec.ts
@@ -1,4 +1,4 @@
-import { shallowWrapperFactory, wrapperFactory } from '../jest-wrapper-factory'
+import { shallowWrapperFactory, wrapperFactory } from '../vitest-wrapper-factory'
 import EntityInfo from '@/components/common/EntityInfo.vue'
 import { FilingTypes } from '@/enums'
 import { CorpTypeCd } from '@bcrs-shared-components/enums/'
