+++ tests/unit/RestorationReviewConfirm.spec.ts
@@ -1,4 +1,4 @@
-import { shallowWrapperFactory } from '../jest-wrapper-factory'
+import { shallowWrapperFactory } from '../vitest-wrapper-factory'
 import { RestorationReviewConfirm } from '@/views'
 import { RestorationResources } from '@/resources/'
 import CardHeader from '@/components/common/CardHeader.vue'
